Backtracking/WordSearch/WordSearch_test.go
package wordSearch

import (
	"fmt"
	"testing"

	"github.com/stretchr/testify/assert"
)

func Test(t *testing.T) {
	testTable := []struct {
		board  [][]byte
		word   string
		output bool
	}{
		{
			board:  [][]byte{{'A', 'B', 'C', 'E'}, {'S', 'F', 'C', 'S'}, {'A', 'D', 'E', 'E'}},
			word:   "ABCCED",
			output: true,
		},
		{
			board:  [][]byte{{'A', 'B', 'C', 'E'}, {'S', 'F', 'C', 'S'}, {'A', 'D', 'E', 'E'}},
			word:   "SEE",
			output: true,
		},
		{
			board:  [][]byte{{'A', 'B', 'C', 'E'}, {'S', 'F', 'C', 'S'}, {'A', 'D', 'E', 'E'}},
			word:   "ABCB",
			output: false,
		},
	}
	for i, testCase := range testTable {
		t.Run(fmt.Sprintf("Example №%v", i+1), func(t *testing.T) {
			assert.Equal(t, testCase.output, exist(testCase.board, testCase.word))
		})
	}
}
